+++ docs/examples/plot_SDSS_r.py
@@ -2,7 +2,7 @@ import astropy.units as u
 import matplotlib.pyplot as plt
 from astropy.coordinates import Angle, SkyCoord
 
-from mocpy import MOC, World2ScreenMPL
+from mocpy import MOC, WCS
 
 # Load a MOC
 filename = "./../../resources/P-SDSS9-r.fits"
@@ -10,18 +10,11 @@ moc = MOC.from_fits(filename)
 # Plot the MOC using matplotlib
 fig = plt.figure(111, figsize=(15, 10))
 # Define a astropy WCS easily
-with World2ScreenMPL(
-    fig,
-    fov=200 * u.deg,
-    center=SkyCoord(0, 20, unit="deg", frame="icrs"),
-    coordsys="icrs",
-    rotation=Angle(0, u.degree),
-    projection="AIT",
-) as wcs:
-    ax = fig.add_subplot(1, 1, 1, projection=wcs)
-    # Call fill with a matplotlib axe and the `~astropy.wcs.WCS` wcs object.
-    moc.fill(ax=ax, wcs=wcs, alpha=0.5, fill=True, color="green")
-    moc.border(ax=ax, wcs=wcs, alpha=0.5, color="black")
+wcs = moc.get_wcs(fig, coordsys="icrs", rotation=Angle(0, u.degree), projection="AIT")
+ax = fig.add_subplot(1, 1, 1, projection=wcs)
+# Call fill with a matplotlib axe and the `~astropy.wcs.WCS` wcs object.
+moc.fill(ax=ax, wcs=wcs, alpha=0.5, fill=True, color="green")
+moc.border(ax=ax, wcs=wcs, alpha=0.5, color="black")
 plt.xlabel("ra")
 plt.ylabel("dec")
 plt.title("Coverage of P-SDSS9-r")